util/metric: emit evicted label in high cardinality metrics
Previously, we were emitting metrics with aggregated value and all child
metrics. This was inadequate because it is hard to keep the track of all
evicted entries. This patch introduces additional label `value_type`. For
parent metrics, the value is `aggregated`. All values against evicted entries
are combined against the value `evicted`.

Epic: None
Part of: CRDB-54635
Release note: None